requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1  # Si decides usar un archivo .env para el TOKEN